Move AppConfig and monkey-patching to apps.py
